requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.1
httpx==0.25.2
websockets==12.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scipy==1.11.4
joblib==1.3.2